move NavigationMapView to own file
