+++ MapboxNavigationUI/MGLMapView.swift
@@ -177,18 +177,3 @@ extension MGLMapView {
 protocol NavigationMapViewDelegate {
     func navigationMapView(_ mapView: NavigationMapView, shouldUpdateTo location: CLLocation) -> CLLocation?
 }
-
-public class NavigationMapView: MGLMapView {
-    
-    var navigationMapDelegate: NavigationMapViewDelegate?
-    
-    public override func locationManager(_ manager: CLLocationManager!, didUpdateLocations locations: [Any]!) {
-        guard let location = locations.first as? CLLocation else { return }
-        
-        if let modifiedLocation = navigationMapDelegate?.navigationMapView(self, shouldUpdateTo: location) {
-            super.locationManager(manager, didUpdateLocations: [modifiedLocation])
-        } else {
-            super.locationManager(manager, didUpdateLocations: locations)
-        }
-    }
-}
